Utils removed + refactor
