+++ package.json
@@ -1,27 +1,25 @@
 {
   "dependencies": {
-    "@distube/ytdl-core": "^4.14.4",
     "axios": "^1.7.4",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
-    "express": "^4.19.2",
+    "express": "^4.21.0",
     "ffmpeg": "^0.0.4",
     "ffmpeg-static": "^5.2.0",
-    "http-proxy": "^1.18.1",
-    "http-proxy-agent": "^7.0.2",
-    "https": "^1.0.0",
-    "https-proxy-agent": "^7.0.5",
+    "fluent-ffmpeg": "^2.1.3",
+    "got": "^11.8.6",
     "morgan": "^1.10.0",
+    "node-fetch": "^2.7.0",
     "nodemon": "^3.1.1",
+    "play-dl": "^1.9.7",
     "readline": "^1.3.0",
-    "socket-io": "^1.0.0",
     "socket.io": "^4.7.5",
     "soundcloud-downloader": "^1.0.0",
     "spotify-uri": "^4.1.0",
     "spotify-web-api-node": "^5.0.2",
     "spotifydl-core": "^0.2.0",
-    "undici": "^6.19.8",
-    "yt-search": "^2.12.1"
+    "tough-cookie": "^5.0.0",
+    "undici": "^6.19.8"
   },
   "name": "media-download-api",
   "description": "An API I've created originally to use for my link utility site",
